save-analysis: use a decoupled representation for dumped data
Closes #33348
