+++ src/librustc_save_analysis/csv_dumper.rs
@@ -10,25 +10,20 @@
 
 use std::io::Write;
 
-use rustc::hir::def_id::{DefId, DefIndex};
-use syntax::codemap::Span;
-
-use super::data::*;
+use super::external_data::*;
 use super::dump::Dump;
-use super::span_utils::SpanUtils;
 
-pub struct CsvDumper<'tcx, 'b, W: 'b> {
-    output: &'b mut W,
-    span: SpanUtils<'tcx>
+pub struct CsvDumper<'b, W: 'b> {
+    output: &'b mut W
 }
 
-impl<'a, 'b, W: Write> CsvDumper<'a, 'b, W> {
-    pub fn new(writer: &'b mut W, span: SpanUtils<'a>) -> CsvDumper<'a, 'b, W> {
-        CsvDumper { output: writer, span: span }
+impl<'b, W: Write> CsvDumper<'b, W> {
+    pub fn new(writer: &'b mut W) -> CsvDumper<'b, W> {
+        CsvDumper { output: writer }
     }
 
-    fn record(&mut self, kind: &str, span: Span, values: String) {
-        let span_str = self.span.extent_str(span);
+    fn record(&mut self, kind: &str, span: SpanData, values: String) {
+        let span_str = span_extent_str(span);
         if let Err(_) = write!(self.output, "{},{}{}\n", kind, span_str, values) {
             error!("Error writing output");
         }
@@ -41,7 +36,7 @@ impl<'a, 'b, W: Write> CsvDumper<'a, 'b, W> {
     }
 }
 
-impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
+impl<'b, W: Write + 'b> Dump for CsvDumper<'b, W> {
     fn crate_prelude(&mut self, data: CratePreludeData) {
         let values = make_values_str(&[
             ("name", &data.crate_name),
@@ -93,21 +88,14 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn impl_data(&mut self, data: ImplData) {
-        let self_ref = data.self_ref.unwrap_or(null_def_id());
-        let trait_ref = data.trait_ref.unwrap_or(null_def_id());
-
         let id = data.id.to_string();
-        let ref_id = self_ref.index.as_usize().to_string();
-        let ref_id_crate = self_ref.krate.to_string();
-        let trait_id = trait_ref.index.as_usize().to_string();
-        let trait_id_crate = trait_ref.krate.to_string();
+        let ref_id = data.self_ref.unwrap_or(Id::null()).to_string();
+        let trait_id = data.trait_ref.unwrap_or(Id::null()).to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("id", &id),
             ("refid", &ref_id),
-            ("refidcrate", &ref_id_crate),
             ("traitid", &trait_id),
-            ("traitidcrate", &trait_id_crate),
             ("scopeid", &scope)
         ]);
 
@@ -115,33 +103,24 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn inheritance(&mut self, data: InheritanceData) {
-       let base_id = data.base_id.index.as_usize().to_string();
-       let base_crate = data.base_id.krate.to_string();
-       let deriv_id = data.deriv_id.to_string();
-       let deriv_crate = 0.to_string();
-       let values = make_values_str(&[
-           ("base", &base_id),
-           ("basecrate", &base_crate),
-           ("derived", &deriv_id),
-           ("derivedcrate", &deriv_crate)
-       ]);
+        let base_id = data.base_id.to_string();
+        let deriv_id = data.deriv_id.to_string();
+        let values = make_values_str(&[
+            ("base", &base_id),
+            ("derived", &deriv_id),
+        ]);
 
        self.record("inheritance", data.span, values);
     }
 
     fn function(&mut self, data: FunctionData) {
-        let (decl_id, decl_crate) = match data.declaration {
-            Some(id) => (id.index.as_usize().to_string(), id.krate.to_string()),
-            None => (String::new(), String::new())
-        };
-
         let id = data.id.to_string();
+        let decl_id = data.declaration.unwrap_or(Id::null()).to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("id", &id),
             ("qualname", &data.qualname),
             ("declid", &decl_id),
-            ("declidcrate", &decl_crate),
             ("scopeid", &scope)
         ]);
 
@@ -149,12 +128,10 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn function_ref(&mut self, data: FunctionRefData) {
-        let ref_id = data.ref_id.index.as_usize().to_string();
-        let ref_crate = data.ref_id.krate.to_string();
+        let ref_id = data.ref_id.to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
             ("qualname", ""),
             ("scopeid", &scope)
         ]);
@@ -163,13 +140,11 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn function_call(&mut self, data: FunctionCallData) {
-        let ref_id = data.ref_id.index.as_usize().to_string();
-        let ref_crate = data.ref_id.krate.to_string();
+        let ref_id = data.ref_id.to_string();
         let qualname = String::new();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
             ("qualname", &qualname),
             ("scopeid", &scope)
         ]);
@@ -190,21 +165,12 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn method_call(&mut self, data: MethodCallData) {
-        let (dcn, dck) = match data.decl_id {
-            Some(declid) => (declid.index.as_usize().to_string(), declid.krate.to_string()),
-            None => (String::new(), String::new()),
-        };
-
-        let ref_id = data.ref_id.unwrap_or(null_def_id());
-
-        let def_id = ref_id.index.as_usize().to_string();
-        let def_crate = ref_id.krate.to_string();
+        let decl_id = data.decl_id.unwrap_or(Id::null()).to_string();
+        let ref_id = data.ref_id.unwrap_or(Id::null()).to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
-            ("refid", &def_id),
-            ("refidcrate", &def_crate),
-            ("declid", &dcn),
-            ("declidcrate", &dck),
+            ("refid", &ref_id),
+            ("declid", &decl_id),
             ("scopeid", &scope)
         ]);
 
@@ -245,15 +211,11 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn mod_ref(&mut self, data: ModRefData) {
-        let (ref_id, ref_crate) = match data.ref_id {
-            Some(rid) => (rid.index.as_usize().to_string(), rid.krate.to_string()),
-            None => (0.to_string(), 0.to_string())
-        };
+        let ref_id = data.ref_id.unwrap_or(Id::null()).to_string();
 
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
             ("qualname", &data.qualname),
             ("scopeid", &scope)
         ]);
@@ -320,15 +282,10 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn type_ref(&mut self, data: TypeRefData) {
-        let (ref_id, ref_crate) = match data.ref_id {
-            Some(id) => (id.index.as_usize().to_string(), id.krate.to_string()),
-            None => (0.to_string(), 0.to_string())
-        };
-
+        let ref_id = data.ref_id.unwrap_or(Id::null()).to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
             ("qualname", &data.qualname),
             ("scopeid", &scope)
         ]);
@@ -348,16 +305,12 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn use_data(&mut self, data: UseData) {
-        let mod_id = data.mod_id.unwrap_or(null_def_id());
-
         let id = data.id.to_string();
-        let ref_id = mod_id.index.as_usize().to_string();
-        let ref_crate = mod_id.krate.to_string();
+        let mod_id = data.mod_id.unwrap_or(Id::null()).to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
             ("id", &id),
-            ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
+            ("mod_id", &mod_id),
             ("name", &data.name),
             ("scopeid", &scope)
         ]);
@@ -395,12 +348,10 @@ impl<'a, 'b, W: Write + 'b> Dump for CsvDumper<'a, 'b, W> {
     }
 
     fn variable_ref(&mut self, data: VariableRefData) {
-        let ref_id = data.ref_id.index.as_usize().to_string();
-        let ref_crate = data.ref_id.krate.to_string();
+        let id = data.ref_id.to_string();
         let scope = data.scope.to_string();
         let values = make_values_str(&[
-            ("refid", &ref_id),
-            ("refidcrate", &ref_crate),
+            ("id", &id),
             ("qualname", ""),
             ("scopeid", &scope)
         ]);
@@ -431,9 +382,9 @@ fn make_values_str(pairs: &[(&'static str, &str)]) -> String {
     })
 }
 
-fn null_def_id() -> DefId {
-    DefId {
-        krate: 0,
-        index: DefIndex::new(0),
-    }
+fn span_extent_str(span: SpanData) -> String {
+    format!("file_name,\"{}\",file_line,{},file_col,{},byte_start,{}\
+             file_line_end,{},file_col_end,{},byte_end,{}",
+             span.file_name, span.line_start, span.column_start, span.byte_start,
+             span.line_end, span.column_end, span.byte_end)
 }
